mod file

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -3,7 +3,7 @@
 /**
  * set_bit - sets the value of a bit to 1 at a given index.
  *
- * @n: unsigned long int 
+ * @n: unsigned long int
  * @index: index to be set
  *
  * Return: 1 if successfull, -1 if an error occured
@@ -11,7 +11,7 @@
 
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	long unsigned int mask;
+	int mask;
 
 	if (index >= sizeof(unsigned long int) * 8)
 		return (-1);
